refactor(desktop): 提取会话快照与子代理状态视图
